Prepare EditorUI for UI text wrapping - Refactor UIEditorPropertiesPanel.cpp/.h to seperate Text Widget code from general widget code - Add editor UI for modifying whether a text widget should be wrapped or not in UIEditorPropertiesPanel.cpp/.h - Fix warnings in UIEditorViewportPanel.cpp - Add m_TextWrapped boolean to RuntimeUI.h/UserInterfaceManager.cpp serialization/deserialization

+++ Editor/Source/Windows/UserInterface/UIEditorViewportPanel.cpp
@@ -653,7 +653,7 @@ namespace Kargono::Panels
 				}
 				else
 				{
-					widget->m_PixelPosition.x = screenPosition.x;
+					widget->m_PixelPosition.x = (int)screenPosition.x;
 				}
 
 				// Update the widget's y position
@@ -663,7 +663,7 @@ namespace Kargono::Panels
 				}
 				else
 				{
-					widget->m_PixelPosition.y = screenPosition.y;
+					widget->m_PixelPosition.y = (int)screenPosition.y;
 				}
 				
 			}
